requirements.txt
Django==3.2.20
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
pymongo==3.12.3
python-decouple==3.8
django-cors-headers==4.2.0
celery==5.3.1
redis==4.6.0
faker==19.3.0
coverage==7.2.7
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0